Moved blockChannel functionality to the Kids client and updated API docs.
