+++ src/core/clients/Kids.ts
@@ -1,15 +1,19 @@
+import Parser from '../../parser/index.js';
 import Channel from '../../parser/ytkids/Channel.js';
 import HomeFeed from '../../parser/ytkids/HomeFeed.js';
 import Search from '../../parser/ytkids/Search.js';
 import VideoInfo from '../../parser/ytkids/VideoInfo.js';
 import type Session from '../Session.js';
+import type { ApiResponse } from '../Actions.js';
 
-import { generateRandomString } from '../../utils/Utils.js';
+import { InnertubeError, generateRandomString } from '../../utils/Utils.js';
 
 import {
   BrowseEndpoint, NextEndpoint,
   PlayerEndpoint, SearchEndpoint
 } from '../endpoints/index.js';
+import { KidsBlocklistPickerEndpoint } from '../endpoints/channel/index.js';
+import KidsBlocklistPickerItem from '../../parser/classes/ytkids/KidsBlocklistPickerItem.js';
 
 export default class Kids {
   #session: Session;
@@ -80,4 +84,37 @@ export default class Kids {
     );
     return new HomeFeed(this.#session.actions, response);
   }
+
+    /**
+   * Retrieves the list of supervised accounts that the signed-in user has 
+   * access to, and blocks the given channel for each of them.
+   * @param channel_id - The channel id to block.
+   * @returns A list of API responses.
+   */
+    async blockChannel(channel_id: string): Promise<ApiResponse[]> {
+      if (!this.#session.logged_in)
+        throw new InnertubeError('You must be signed in to perform this operation.');
+  
+      const blocklist_payload = KidsBlocklistPickerEndpoint.build({ channel_id: channel_id });
+      const response = (await this.#session.actions.execute( KidsBlocklistPickerEndpoint.PATH, blocklist_payload ));     
+      const popup = response.data.command.confirmDialogEndpoint    
+      const popupFragment = { contents: popup.content, engagementPanels: [] };
+      const kidPicker = Parser.parseResponse(popupFragment);
+      const kids = kidPicker.contents_memo?.getType(KidsBlocklistPickerItem)
+      
+      if (!kids)
+        throw new InnertubeError('Could not find any kids profiles or supervised accounts.');
+  
+      // Iterate through the kids and block the channel if not already blocked.
+      var responses: ApiResponse[] = [];
+      for (const kid of kids) {
+        if(!kid.block_button?.is_toggled) {            
+          kid.setActions(this.#session.actions);
+          // Block channel and add to the response list. 
+          responses.push(await kid.blockchannel());     
+        }
+      }
+  
+      return responses;
+    }
 }
