+++ src/core/managers/InteractionManager.ts
@@ -2,14 +2,11 @@ import * as Proto from '../../proto/index.js';
 import type Actions from '../Actions.js';
 import type { ApiResponse } from '../Actions.js';
 
-import { InnertubeError, throwIfMissing } from '../../utils/Utils.js';
+import { throwIfMissing } from '../../utils/Utils.js';
 import { LikeEndpoint, DislikeEndpoint, RemoveLikeEndpoint } from '../endpoints/like/index.js';
 import { SubscribeEndpoint, UnsubscribeEndpoint } from '../endpoints/subscription/index.js';
 import { CreateCommentEndpoint, PerformCommentActionEndpoint } from '../endpoints/comment/index.js';
 import { ModifyChannelPreferenceEndpoint } from '../endpoints/notification/index.js';
-import { KidsBlocklistPickerEndpoint } from '../endpoints/channel/index.js';
-import KidsBlocklistPickerItem from '../../parser/classes/channels/KidsBlocklistPickerItem.js';
-import Parser from '../../parser/index.js';
 
 export default class InteractionManager {
   #actions: Actions;
@@ -199,38 +196,5 @@ export default class InteractionManager {
     );
 
     return action;
-  }
-
-  /**
-   * Retrieves the list of supervised accounts that the signed-in user has 
-   * access to, and blocks the given channel for each of them.
-   * @param channel_id - The channel id to block.
-   * @returns A list of API responses.
-   */
-  async blockChannel(channel_id: string): Promise<ApiResponse[]> {
-    if (!this.#actions.session.logged_in)
-      throw new InnertubeError('You must be signed in to perform this operation.');
-
-    const blocklist_payload = KidsBlocklistPickerEndpoint.build({ channel_id: channel_id });
-    const response = (await this.#actions.execute( KidsBlocklistPickerEndpoint.PATH, blocklist_payload ));     
-    const popup = response.data.command.confirmDialogEndpoint    
-    const popupFragment = { contents: popup.content, engagementPanels: [] };
-    const kidPicker = Parser.parseResponse(popupFragment);
-    const kids = kidPicker.contents_memo?.getType(KidsBlocklistPickerItem)
-    
-    if (!kids)
-      throw new InnertubeError('Could not find any kids profiles or supervised accounts.');
-
-    // Iterate through the kids and block the channel if not already blocked.
-    var responses: ApiResponse[] = [];
-    for (const kid of kids) {
-      if(!kid.block_button?.is_toggled) {            
-        kid.setActions(this.#actions);
-        // Block channel and add to the response list. 
-        responses.push(await kid.blockchannel());     
-      }
-    }
-
-    return responses;
   }
 }
